Fixes edge sizes in main edge program
There were multiple issues with edge sizes before:
- Edges were not rendered with the same size in Retina screens and
non-Retina screens
- Setting an edge size in pixels in the data would not fit the pixels on
screen, at any point

The core reason of this is that all the adjustments we added to the
matrix to handle rescaling messed up all the measurements of sizes in
the WebGL shaders.

To fix this, I added a way to "counter" the impact of this matrix, with
a flat `correctionRatio` given to every programs. This ratio is useless
when using GL_POINTS or GL_LINES, but can fix all those size issues in
every other programs.

TODO:
- Fix anti-aliasing
- Fix remaining relevant programs
- Remove dead programs (node.ts looks like it does not work anymore)
- Fix e2e tests